LayoutTests/fast/forms/number/number-input-switch-from-text-clears-invalid-number-value-expected.txt

Verify that switching input from type=text to number drops invalid values but preserves valid numeric values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Switching type=text to type=number for value="abc"
PASS input.value is ""
Switching type=text to type=number for value="--123"
PASS input.value is ""
Switching type=text to type=number for value="１２３"
PASS input.value is ""
Switching type=text to type=number for value="1e--2"
PASS input.value is ""
Switching type=text to type=number for value="123"
PASS input.value is "123"
Switching type=text to type=number for value="+123"
PASS input.value is ""
Switching type=text to type=number for value="-123.45"
PASS input.value is "-123.45"
Switching type=text to type=number for value="1e5"
PASS input.value is "1e5"
PASS successfullyParsed is true

TEST COMPLETE
